feat: divides the code into modules, connects the modules, displays a function getPhotosDataArray in main.js
