+++ index.html
@@ -251,8 +251,8 @@
     </section>
   </template>
 
-  <script src="js/functions.js"></script>
-  <script src="js/main.js"></script>
+  <script src="js/main.js" type="module"></script>
+
 </body>
 
 </html>
