+++ js/functions.js
@@ -57,3 +57,4 @@ getNumbers('0029 of June 2024 year');
 // имяФункции(2023); // 2023
 // имяФункции(-1);   // 1
 // имяФункции(1.5);  // 15
+
